Styled the SelectDifficulty page in a similar way to WelcomePage
Noticed a bug, whenever I click pause in DisplayGame.jsx, it pauses and then after
some seconds it continues. I commented out the timer feature I wrote to change the animals displaying
in WelcomePage and SelectDifficulty page and the bug ceased. I don't know the reason but I'd come back to it later.